+++ react/Footer.js
@@ -1,3 +1,0 @@
-import Footer from './components/Footer/index'
-
-export default Footer